community/dai-complex-workflow/compile_hitl_results/requirements.txt
google-api-core==1.28.0
google-auth==1.30.0
google-cloud-core==1.6.0
google-cloud-documentai==0.5.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-cloud-bigquery==2.17.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
idna==2.10
pandas==1.3.2
packaging==20.9
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
PyPDF2==1.26.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
urllib3==1.26.4
fuzzy-match==0.0.1